Extract acceptance of snapshot and creation of context
